These updates should get master stable again

+++ FirstProject/src/RenderText.cpp
@@ -8,7 +8,7 @@ RenderText::RenderText(int screen_width, int screen_height)
 	this->projection_transform = glm::ortho(0.0f, static_cast<GLfloat>(screen_width), 0.0f, static_cast<GLfloat>(screen_height));
 	glUseProgram(this->program);
 	glUniformMatrix4fv(glGetUniformLocation(this->program, "projection"), 1, GL_FALSE, glm::value_ptr(this->projection_transform));
-	
+
 	// FreeType
 	FT_Library ft;
 	// All functions return a value different than 0 whenever an error occurred
@@ -64,7 +64,7 @@ RenderText::RenderText(int screen_width, int screen_height)
 			texture,
 			glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 			glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-			face->glyph->advance.x
+			static_cast<GLuint>(face->glyph->advance.x)
 		};
 		this->Characters.insert(std::pair<GLchar, Character>(c, character));
 	}
@@ -130,4 +130,3 @@ void RenderText::DrawString(std::string text, GLfloat x, GLfloat y, GLfloat scal
 	glBindVertexArray(0);
 	glBindTexture(GL_TEXTURE_2D, 0);
 }
-
